app/tools/Makefile.am added gimp_tool_motion_constrain(),
2004-08-22  Sven Neumann  <sven@gimp.org>

	* app/tools/Makefile.am
	* app/tools/tools-utils.[ch]: added gimp_tool_motion_constrain(),

	* app/paint/gimppaintcore.[ch]: removed gimp_paint_core_constrain().

	* app/tools/gimppainttool.c: changed accordingly.

	* app/tools/gimpblendtool.[ch]: use gimp_tool_motion_constrain()
	instead of duplicating that functionality.

	* app/tools/gimpmeasuretool.c: use gimp_tool_motion_constrain()
	instead of implementing completely different constraints.

+++ app/tools/gimpblendtool.h
@@ -38,10 +38,10 @@ struct _GimpBlendTool
 {
   GimpDrawTool  parent_instance;
 
-  gint          startx;      /*  starting x coord     */
-  gint          starty;      /*  starting y coord     */
-  gint          endx;        /*  ending x coord       */
-  gint          endy;        /*  ending y coord       */
+  gdouble       startx;      /*  starting x coord     */
+  gdouble       starty;      /*  starting y coord     */
+  gdouble       endx;        /*  ending x coord       */
+  gdouble       endy;        /*  ending y coord       */
 };
 
 struct _GimpBlendToolClass
